Separate begingroup into three files, as requested
